calculator backup

+++ RFA.ENTITY/Dto/PaymentInfoDtos/ResponseDtos/PaymentCalculatorResponseDto.cs
@@ -9,11 +9,11 @@ namespace RFA.ENTITY.Dto.PaymentInfoDtos.ResponseDtos
 {
     public class PaymentCalculatorResponseDto : IDto
     {
-        public int? ParticipationFee { get; set; }
-        public int? GalaFee { get; set; }
-        public int? AccommodationFee { get; set; }
-        public int? CourseFee { get; set; }
-        public int? Discount { get; set; }
-        public int? Total { get; set; }
+        public double? ParticipationFee { get; set; }
+        public double? GalaFee { get; set; }
+        public double? AccommodationFee { get; set; }
+        public double? CourseFee { get; set; }
+        public double? Discount { get; set; }
+        public double? Total { get; set; }
     }
 }
